ADDED Modal pages for settings, CHANGED session page to reflect design ADDED links to setting on everypage
I added the last pages and added the links to those pages. I changed the
way session worked by faking it (link to other page). I animated the
modal pages but linked it to an other page because 2 modals on 1 page
didn't work.And lastly I added some images that were needed to finish
the pages